\name{simone-internal}

\alias{analysis}
\alias{color.lightening}
\alias{CoopLassoConstraint}
\alias{dL1}
\alias{dL2}
\alias{dL3}
\alias{draw.arc}
\alias{draw.edges}
\alias{draw.circle}
\alias{draw.sector}
\alias{draw.loops}
\alias{draw.rot.poly}
\alias{get.rho.max}
\alias{getDegree}
\alias{Gplot}
\alias{Gplot.add}
\alias{Gplot.default}
\alias{Gplot.delete}
\alias{Gplot.edge}
\alias{Gplot.graphics}
\alias{Gplot.inference}
\alias{Gplot.network}
\alias{Gplot.pie.nodes}
\alias{Gplot.radial.vertex.label}
\alias{Gplot.vertex}
\alias{Gplot.vertex.label}
\alias{Gplot.vertex.pie}
\alias{graphical.lasso}
\alias{GroupLassoConstraint}
\alias{InferEdges}
\alias{InferStructure}
\alias{is.simone}
\alias{is.simone.network}
\alias{is.symetric}
\alias{isolate.vertices}
\alias{L1}
\alias{L2}
\alias{L3}
\alias{laplacian}
\alias{LassoConstraint}
\alias{light.palette}
\alias{multiple.dynamic}
\alias{multiple.static}
\alias{nb.edges}
\alias{neighborhood.selection}
\alias{OptInference}
\alias{plot.path}
\alias{plot.BIC}
\alias{plot.AIC}
\alias{plot.ROC}
\alias{plot.PR}
\alias{plot.sequence}
\alias{sub.param}
\alias{var1.inference}

\title{'simone' internal functions}
\description{
  Internal functions.
}

\details{
  SIMoNe internal functions: not to be called by the user.
}

\author{C. Ambroise, C. Charbonnier, J. Chiquet, G. Grasseau, A. Smith}

\keyword{internal}