\name{simone-internal}

\alias{BaseLambdaValue}
\alias{CalcDegrees}
\alias{CheckClassLoss}
\alias{ClassNum2Indic}
\alias{draw.edges}
\alias{get.label.index.list}
\alias{Gplot.edge} 
\alias{Gplot.graphics}
\alias{Gplot.network}
\alias{Gplot.vertex}
\alias{Gplot.vertex.label}
\alias{grey.palette}
\alias{IC.Estep}
\alias{IC.Mstep}
\alias{isolate.vertices}
\alias{JRx}
\alias{light.palette}
\alias{lnFLaplace}
\alias{Mplot.env}
\alias{SimConnectivity}
\alias{SimGaussianVector}
\alias{SimPrecMat}
\alias{SpectralClustering}
\alias{sub.param}
\alias{ticks.space}
\alias{weighted.median}

\title{'simone' internal functions}
\description{
  Internal functions.
}

\details{
  Not to be called by the user.
}

\author{C. Ambroise, J. Chiquet, G. Grasseau, A. Smith}

\keyword{internal}